requirements.txt
pysqlite3-binary
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.13
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
deprecated==1.2.14
distro==1.9.0
durationpy==0.9
effdet==0.4.1
emoji==2.14.0
eval-type-backport==0.2.0
fastapi==0.115.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.4.0
importlib-resources==6.4.5
iopath==0.1.10
jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
langdetect==1.0.9
layoutparser==0.3.4
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.9.2
nest-asyncio==1.6.0
networkx==3.4
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.51.2
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pdfminer-six==20240706
pdfplumber==0.5.3
pi-heif==0.18.0
pikepdf==9.3.0
pillow==10.4.0
portalocker==2.10.1
posthog==3.7.0
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycrypto==2.6.1
pydantic==2.9.2
pydantic-core==2.23.4
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.4
pypdf==5.0.1
pypdf2==3.0.1
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.9
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
streamlit==1.39.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
timm==1.0.9
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unicodecsv==0.14.1
unstructured==0.15.14
unstructured-client==0.26.0
unstructured-inference==0.7.36
unstructured-pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
wand==0.6.13
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wordcloud==1.9.3
wrapt==1.16.0
zipp==3.20.2
